fixed compiler warnings
git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/trunk@2837 494c09c7-a67d-49d4-9938-151fb1965be8

+++ RobWork/src/rw/geometry/Cone.cpp
@@ -52,7 +52,7 @@ TriMesh::Ptr Cone::createMesh(int resolution) const{
 
 	PlainTriMeshF *mesh = new PlainTriMeshF(4*level);
 
-	float z = _height/2.0f;
+	float z = (float)(_height/2.0f);
 
 	for (int i = 0; i < level; i++) {
 		//Construct Triangles for curved surface
